EngineResult with Meta

+++ src/algorithm/metric.rs
@@ -1,12 +1,12 @@
 pub trait Metric {
     fn get_nodes(&self) -> usize;
     fn add_nodes(&mut self, increment: usize);
-    fn get_depth(&self) -> usize;
-    fn put_depth(&mut self, depth: usize);
+    fn get_depth(&self) -> u8;
+    fn put_depth(&mut self, depth: u8);
 }
 
 pub struct Meta {
-    depth: usize,
+    depth: u8,
     nodes: usize,
 }
 
@@ -23,10 +23,10 @@ impl Metric for Meta {
     fn add_nodes(&mut self, increment: usize) {
         self.nodes = self.nodes + increment
     }
-    fn get_depth(&self) -> usize {
+    fn get_depth(&self) -> u8 {
         self.depth
     }
-    fn put_depth(&mut self, depth: usize) {
+    fn put_depth(&mut self, depth: u8) {
         if self.depth < depth {
             self.depth = depth
         }
